advanced_api_project/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.5
